境外服务器无法获取歌曲时长的问题；歌曲播放完毕的判断问题

+++ tools-os/src/main/java/me/znzn/tools/module/music/entity/MusicPushVO.java
@@ -36,7 +36,7 @@ public class MusicPushVO extends MessageVO implements Serializable {
 
     private String current;
 
-    private Integer length;
+    private Double length;
 
     @Override
     public String toString() {
